+++ classes/Overview/class.xaliOverviewGUI.php
@@ -23,6 +23,7 @@ class xaliOverviewGUI extends xaliGUI {
 	const CMD_SAVE_ENTRY = 'saveEntry';
 	const CMD_SAVE_USER = 'saveUser';
     const CMD_ADD_USER_AUTO_COMPLETE = 'addUserAutoComplete';
+	const CMD_SAVE_ABSENCE_REASON = 'saveAbsenceReason';
 
 	const SUBTAB_USERS = 'subtab_users';
 	const SUBTAB_LISTS = 'subtab_lists';
@@ -384,5 +385,28 @@ class xaliOverviewGUI extends xaliGUI {
         $list = json_encode($array);
         echo $list;
         exit();
+    }
+
+
+    /**
+     *
+     */
+    protected function saveAbsenceReason()
+    {
+        /** @var xaliChecklist $checklist */
+        $checklist = xaliChecklist::find($_GET['checklist_id']);
+
+        $entry = $checklist->getEntryOfUser($_GET['user_id']);
+
+        if (intval($entry->getStatus()) === xaliChecklistEntry::STATUS_ABSENT_UNEXCUSED) {
+            if (($reason_id = $_GET['absence_reason']) !== null) {
+                /** @var xaliAbsenceStatement $stm */
+                $stm = xaliAbsenceStatement::findOrGetInstance($entry->getId());
+                $stm->setReasonId($reason_id);
+                $stm->store();
+            }
+        }
+
+        exit;
     }
 }